Move screen scheduling tests to new file
